Check if snapshot exists
Added an if statement to check if a snapshot exists when searching for an item and when an item is clicked on the page to be deleted. The else statement displays text that says no items are in the shopping list currently.
Added hover effects for shopping list items and the add to cart button, to change colour and change cursor to a mouse pointer.